TeamB/tamanna_requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
autoawq==0.2.9
bitsandbytes==0.49.2
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.5.4
cuda-toolkit==12.8.1
cycler==0.12.1
datasets==4.8.5
defuser==0.0.21
device-smi==0.5.6
dill==0.4.1
filelock==3.29.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
gitpython==3.1.49
gptqmodel==7.0.0
gpytorch==1.15.2
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.12.2
idna==3.13
jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
linear-operator==0.6.1
logbar==0.4.3
markdown==3.10.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.9
maturin==1.13.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
narwhals==2.20.0
netcal==1.3.5
networkx==3.6.1
ninja==1.13.0
numpy==2.2.6
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.20.0.48
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile==1.15.1.6
nvidia-cufile-cu12==1.13.1.3
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.29.7
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.8.90
opt-einsum==3.4.0
optimum==2.1.0
packaging==26.2
pandas==3.0.2
pillow==12.2.0
platformdirs==4.9.6
plotly==6.7.0
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pyarrow==24.0.0
pydantic==2.13.3
pydantic-core==2.46.3
pygments==2.20.0
pyparsing==3.3.2
pypcre==0.3.2
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyyaml==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
sentry-sdk==2.58.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tikzplotlib==0.9.8
tokenicer==0.0.13
tokenizers==0.22.2
torch==2.12.0.dev20260407+cu128
torchvision==0.27.0.dev20260407+cu128
tqdm==4.67.3
transformers==5.7.0
triton==3.7.0+git9c288bc5
typer==0.25.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
wandb==0.26.1
werkzeug==3.1.8
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
